Fix stm32g4 ADC RCC SPI

+++ lib/stm32/g4/adc.c
@@ -503,19 +503,19 @@ void adc_set_injected_offset(uint32_t adc, uint8_t reg, uint32_t offset)
 {
 	switch (reg) {
 	case 1:
-		ADC_OFR1(adc) |= ADC_OFR1_OFFSET1_EN;
+		ADC_OFR1(adc) |= (uint32_t)ADC_OFR1_OFFSET1_EN;
 		ADC_OFR1(adc) |= offset;
 		break;
 	case 2:
-		ADC_OFR2(adc) |= ADC_OFR2_OFFSET2_EN;
+		ADC_OFR2(adc) |= (uint32_t)ADC_OFR2_OFFSET2_EN;
 		ADC_OFR2(adc) |= offset;
 		break;
 	case 3:
-		ADC_OFR3(adc) |= ADC_OFR3_OFFSET3_EN;
+		ADC_OFR3(adc) |= (uint32_t)ADC_OFR3_OFFSET3_EN;
 		ADC_OFR3(adc) |= offset;
 		break;
 	case 4:
-		ADC_OFR4(adc) |= ADC_OFR4_OFFSET4_EN;
+		ADC_OFR4(adc) |= (uint32_t)ADC_OFR4_OFFSET4_EN;
 		ADC_OFR4(adc) |= offset;
 		break;
 	}
@@ -527,8 +527,8 @@ void adc_set_injected_offset(uint32_t adc, uint8_t reg, uint32_t offset)
  * The ADC clock taken from the APB2 clock can be scaled down by 2, 4, 6 or 8.
  *
  * @param adc peripheral of choice @ref adc_reg_base
- * @param[in] source Unsigned int32. Source value for ADC Clock @ref
- * adc_ccr_adcpre
+ * @param[in] source Unsigned int32. Source value for ADC Clock @ref adc_ccr_presc
+ * 
  */
 void adc_set_clk_source(uint32_t adc, uint32_t source)
 {
@@ -542,8 +542,8 @@ void adc_set_clk_source(uint32_t adc, uint32_t source)
  * The ADC clock taken from the APB2 clock can be scaled down by 2, 4, 6 or 8.
  *
  * @param adc peripheral of choice @ref adc_reg_base
- * @param[in] prescale Unsigned int32. Prescale value for ADC Clock @ref
- * adc_ccr_adcpre
+ * @param[in] prescale Unsigned int32. Prescale value for ADC Clock @ref adc_ccr_presc
+ * 
  */
 void adc_set_clk_prescale(uint32_t adc, uint32_t ckmode)
 {
